Rename chartResponses to chartTypes
Makes it easier to make sure both files in frontend and backend correlate.
